+++ test/_setup.js
@@ -1,0 +1,6 @@
+const { JSDOM } = require('jsdom');
+
+const DOM = new JSDOM('');
+
+global.window = DOM.window;
+global.document = DOM.document;